made pingpong_3D_config directory
