+++ frontend/src/pages/pingpong_3D/style.css
@@ -33,68 +33,3 @@
 	border-radius: 8;
 	cursor: pointer;
 }
-
-.pp3d-config {
-	color: #fff;
-	font-size: 18px;
-}
-
-.pp3d-config-row {
-	margin: 15px 0;
-	display: flex;
-	align-items: center;
-	gap: 10px;
-}
-
-.pp3d-config-row--button {
-	justify-content: center;
-	margin-top: 25px;
-}
-
-.pp3d-config label {
-	min-width: 150px;;
-}
-
-.pp3d-config input,
-.pp3d-config select {
-	color: #000;
-	padding: 6px 10px;
-	border-radius: 6px;
-	border: none;
-	background: #f00;
-	font-size: 16px;
-}
-
-.pp3d-config button {
-	padding: 12px 20px;
-	background: #fc0;
-	color: #000;
-	border-radius: 10px;
-	font-weight: bold;
-	font-size: 18px;
-	cursor: pointer;
-	box-shadow: 0 0 12px rgba(255, 204, 0 , 0.7);
-}
-
-.pp3d-config button:hover {
-	background-color: #f00;
-}
-
-.on-overflow {
-	overflow: hidden;;
-}
-
-.pp-preview-container {
-	width: 75%;
-	height: 250px;
-	margin-top: 20px;
-	border: 1px solid #aaa;
-	border-radius: 10px;
-	overflow: hidden;
-}
-
-#previewCanvas3D {
-	width: 100%;
-	height: 100%;
-	background: black;
-}
