+++ frontend/src/router/routes.ts
@@ -2,7 +2,8 @@
 import type { Route } from "@/types/routes";
 import { HomeRoute } from "@/pages/home";
 import { PingPongRoute } from "@/pages/pingpong";
-import { PingPong3DSettingRoute, PingPong3DGameRoute } from "@/pages/pingpong_3D";
+import { PingPong3DGameRoute } from "@/pages/pingpong_3D";
+import { PingPong3DSettingRoute } from "@/pages/pingpong_3D_config";
 import { LoginRoute } from "@/pages/login";
 import { WebSocketRoute } from "@/pages/websocket";
